ser config
