+++ src/serializers/config.rs
@@ -1,0 +1,95 @@
+use std::borrow::Cow;
+
+use pyo3::prelude::*;
+use pyo3::types::{PyDelta, PyDict};
+use pyo3::{intern, PyNativeType};
+
+use crate::build_tools::{py_err, SchemaDict};
+use crate::input::pytimedelta_as_duration;
+
+use super::shared::py_err_se_err;
+
+#[derive(Debug, Clone, Copy)]
+pub(crate) struct SerializationConfig {
+    pub timedelta_mode: TimedeltaMode,
+}
+
+impl SerializationConfig {
+    pub fn from_config(config: Option<&PyDict>) -> PyResult<Self> {
+        let timedelta_mode = TimedeltaMode::from_config(config)?;
+        Ok(Self { timedelta_mode })
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+pub(crate) enum TimedeltaMode {
+    Iso8601,
+    Float,
+}
+
+impl TimedeltaMode {
+    pub fn from_config(config: Option<&PyDict>) -> PyResult<Self> {
+        let raw_mode: Option<&str> = match config {
+            Some(c) => c.get_as::<&str>(intern!(c.py(), "ser_json_timedelta"))?,
+            None => None,
+        };
+        match raw_mode {
+            Some("iso8601") => Ok(Self::Iso8601),
+            Some("float") => Ok(Self::Float),
+            Some(s) => py_err!(
+                "Invalid timedelta serialization mode: `{}`, expected `iso8601` or `float`",
+                s
+            ),
+            None => Ok(Self::Iso8601),
+        }
+    }
+
+    fn total_seconds(py_timedelta: &PyDelta) -> PyResult<&PyAny> {
+        py_timedelta.call_method0(intern!(py_timedelta.py(), "total_seconds"))
+    }
+
+    pub fn timedelta_to_json(&self, py_timedelta: &PyDelta) -> PyResult<PyObject> {
+        let py = py_timedelta.py();
+        match self {
+            Self::Iso8601 => {
+                let d = pytimedelta_as_duration(py_timedelta);
+                Ok(d.to_string().into_py(py))
+            }
+            Self::Float => {
+                let seconds = Self::total_seconds(py_timedelta)?;
+                Ok(seconds.into_py(py))
+            }
+        }
+    }
+
+    pub fn json_key<'py>(&self, py_timedelta: &PyDelta) -> PyResult<Cow<'py, str>> {
+        match self {
+            Self::Iso8601 => {
+                let d = pytimedelta_as_duration(py_timedelta);
+                Ok(d.to_string().into())
+            }
+            Self::Float => {
+                let seconds: f64 = Self::total_seconds(py_timedelta)?.extract()?;
+                Ok(seconds.to_string().into())
+            }
+        }
+    }
+
+    pub fn timedelta_serialize<S: serde::ser::Serializer>(
+        &self,
+        py_timedelta: &PyDelta,
+        serializer: S,
+    ) -> Result<S::Ok, S::Error> {
+        match self {
+            Self::Iso8601 => {
+                let d = pytimedelta_as_duration(py_timedelta);
+                serializer.serialize_str(&d.to_string())
+            }
+            Self::Float => {
+                let seconds = Self::total_seconds(py_timedelta).map_err(py_err_se_err)?;
+                let seconds: f64 = seconds.extract().map_err(py_err_se_err)?;
+                serializer.serialize_f64(seconds)
+            }
+        }
+    }
+}
